Implement SSL file resource creation based on paths
This makes the module easier to use (it's based off some work I did a
while ago, rebased onto the latest code). Provide the module with a path
to your SSL cert and key, and it'll copy those into the Prosody folder
with a file resource and the proper permissions.

Speaking of permissions, the Prosody
[docs](https://prosody.im/doc/certificates#installing_the_certificate)
say the cert and key need to be readable by Prosody (and this means
owned by Prosody). I thought this seemed a bit odd, but this [RPM
bug](https://bugzilla.redhat.com/show_bug.cgi?id=1009272) confirms
them. This does make your key as secure as the Prosody server... but
that will be the case no matter what since it does, of course, require
read-access to perform encryptiong.