Fix user creation if domain contains a dash
Prosody encodes every non-word character[1], but I don't know how to
replicate this behaviour in puppet so I'd just adding the dash which I
happen to need. Without this fix, the user file is created in an
incorrect location and prosody will not use it/tell you that the user
does not exist.

[1] https://hg.prosody.im/trunk/file/f5d88ad24b30/util/datamanager.lua#l54

Signed-off-by: Florian Pritz <bluewind@xinu.at>